fix: cleanup code and fix styling
