+++ client/package.json
@@ -15,6 +15,7 @@
     "firebase-tools": "^9.9.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-google-login": "^5.2.2",
     "react-hook-form": "^6.15.4",
     "react-redux": "^7.2.3",
     "react-scripts": "4.0.3",
