+++ client/src/App.tsx
@@ -2,40 +2,16 @@ import './App.css';
 import { FunctionComponent } from 'react';
 import { Layout } from './components/shared/Layout';
 
-import { auth, provider } from './firebase';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
+import { selectUserName } from './redux/userSlice';
 
-import { setActiveUser, setUserLogout, selectUserName } from './redux/userSlice';
+import { Login } from './pages/Login';
+import { Sites } from './pages/Sites';
 
 const App: FunctionComponent = () => {
-  const dispatch = useDispatch();
-
   const userName = useSelector(selectUserName);
 
-  const handleSignIn = () => {
-    auth.signInWithPopup(provider).then((result) => {
-      dispatch(
-        setActiveUser({
-          userName: result.user?.displayName,
-          userEmail: result.user?.email,
-        }),
-      );
-    });
-  };
-  const handleSignOut = () => {
-    auth
-      .signOut()
-      .then(() => {
-        dispatch(setUserLogout());
-      })
-      .catch((error) => alert(error.message));
-  };
-
-  return (
-    <Layout>
-      {userName ? <button onClick={handleSignOut}>Sign out</button> : <button onClick={handleSignIn}>Sign in</button>}
-    </Layout>
-  );
+  return <Layout>{userName ? <Sites /> : <Login />}</Layout>;
 };
 
 export default App;
